+++ packages/design-system-react/src/components/TextArea/index.ts
@@ -1,0 +1,3 @@
+export { TextAreaResize } from './TextArea.constants';
+export { TextArea } from './TextArea';
+export type { TextAreaProps } from './TextArea.types';